avoid cuhpx to reorder xy2nest and back
cuhpx is not open source yet and can be challenging to install.

the conversions between nest and xy pixel ordering are all vectorized
pytorch code so we don't need cuhpx for that. I have also confirmed that
torch.compile is able to correctly fuse the index conversion and array
loading, so this should be as fast as a hand implementation in cuda.