Allow negative longitude angles when regridding from LatLon grids
Previously, the longitudes were assumed to be in the 0...360 range, and
working with data where longitudes were e.g. in the -180...180 range
meant first converting the angles then rearranging the data such that
the longitudes were in ascending order.

When building the regridder, we now use the minimum value in the source
longitudes as the reference point, then we map the destination
longitudes to the same range. For example, if the minimum value is -180,
we map longitudes to the -180...180 range. If it is 123, we map them to
123...483. Crucially, the source longitudes are not changed in any way,
to ensure they are left in ascending order.

This is useful as it allows working with grid data with different
origins without having to pre-process/rearrange it.

Signed-off-by: Ștefan Săftescu <stefan@2lw.ai>